Safari-609.1.11.1/LayoutTests/fast/files/url-required-arguments-expected.txt

Test required arguments of URL object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS URL.createObjectURL() threw exception TypeError: Not enough arguments.
PASS URL.revokeObjectURL() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
